requirements.txt
# FastAPI and ASGI
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Database - SQLAlchemy 2.0 Async
sqlalchemy==2.0.36
alembic==1.13.2
asyncpg==0.30.0  # PostgreSQL async driver
aiomysql==0.2.0   # MySQL async driver

# Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
cryptography==43.0.1

# Redis for caching and rate limiting
redis==5.2.0
hiredis==3.0.0

# AWS SDK for SSM Parameter Store
boto3==1.35.20

# Environment and configuration
python-dotenv==1.0.1

# Logging and monitoring
structlog==24.2.0

# Testing (optional, for development)
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
